requirements.txt

streamlit
langchain
langchain_community
python-docx
PyMuPDF
openpyxl
json
logging
datetime
time
uuid
duckduckgo-search